refactor internal package for importing code as lib
