+++ helpers/lexer/consts.go
