+++ helpers/lexer/eval.go
