+++ helpers/lexer/example_test.go
