+++ helpers/strategyoperators/combinators_test.go
