+++ helpers/strategyoperators/strategyoperators_test.go
